In Apple link endpoint, handle subscriptions being an object
The Feast app is sending the subscriptions as an object, not a single
item array.
